requirements.txt
gTTS==2.5.4
newspaper3k==0.2.8
nltk==3.9.1
lxml_html_clean==0.4.1
ffmpeg==1.4 # Reproductor de MP3 de Linux